fix updateFloodFillForNewPaths freeze

+++ love/src/modules/image/ImageData.h
@@ -145,7 +145,7 @@ private:
 	void decode(Data *data);
 
 	int floodFillTest(int x, int y, ImageData *paths, const Pixel &p);
-	int floodFillTest2(int x, int y, ImageData *paths, uint8 * pixels);
+	int floodFillTest2(int x, int y, ImageData *paths, int * pixels);
 	void clearPixel(Pixel &p);
 	int internalFloodFill(int x, int y, ImageData *paths, const Pixel &p);
